requirements.txt
blinker==1.4
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.13
pyTelegramBotAPI==3.6.6
PyYAML==3.13
requests==2.20.0
sentry-sdk==0.5.3
six==1.11.0
SQLAlchemy==1.2.12
sqlalchemy-mixins==0.2.3
typing==3.6.6
urllib3==1.24.1
Werkzeug==0.14.1
